medbot/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
aiogram==3.10.0
python-dotenv==1.0.1
redis==5.0.7
openai==1.44.0
pydub==0.25.1
requests==2.32.3
httpx<0.28